Ensemble Number, Documentation
Reading through the user manual I found there was another byte in the variable leader data for rollover. This was required to make it so numbering was correct - would have had a lot of ensembles with the same number. Required changing the data type for number to accommodate this
Added comments laying out which bytes correspond to what data to make it easier to track what's going on