Converting from int to floating
Changed it so that data is saved as properly scaled floating point numbers instead of integers that later need to be scaled. Test.py was created with help from chatgpt to make sure everything stayed the same. Old files are saved as _old in case need to swap back later.